Moved new filter and sort functions in store to helper
